MyProjectDB/requirements.txt
# Python dependencies for database management and FastAPI backend
pymongo
python-dotenv
fastapi
uvicorn